Correct document structure
Hi.
Changed tag "h6" to "p" in the creation date section. When creating a post the heading structure is h1, h6 and h2, which gives validation problems and is not a correct document structure. The style/line-hight slightly changes for the display of the the creation date. I have not adjusted that, since i don't think that it's that big of an visual/design impact.

Greetings Nils